Added "Median Seroprevalence by UN Region" and did a small refactor on how we do graphs in recharts. (#193)
Resolves #160 

The TLDR of this PR is that I made a new graph for "Median
Seroprevalence by UN region" and I made two generalized components we
can use for generating future visualizations to make my life easier now
and in the future lol.

Anyways, that's what most of this PR is. I took the old graphs and made
them use my generalized component and then made a new graph using my new
component so we should be able to stand up visualizations faster after
this is merged. I plan on switching some of the older graphs to my new
generalized component but one thing at a time.

![Peek 2024-02-25
20-45](https://github.com/serotracker/dashboard/assets/86806388/beaff9d7-2ed3-4817-9f3b-bef324ef864a)