Issue 158 redesign header (#181)
This PR addresses #158  and #133 

I update our header and navbar to be more in line with the design here:
[figma
|https://www.figma.com/file/BKtEVRMNkZYh4EbG9T5wp5/PathoTracker?type=design&node-id=310-401&mode=design&t=2fv4Vv0lqvPLoVbP-4]

The colors were inverted based on feedback given during UI review of
liking how the dark color up top looked over how the white header
looked.

A lot of the changes were made on a higher level to the shadcn
components themselves. This was because the likelihood that we were to
use these components in another place are very low, and hence for this
use case making those changes up top was great

TODO:

- [x] Make the Heading adapt to the tracker. (Sero -> SC2 -> Arbo)
- [x] Extract the styling applied to the ListItem component as well as
extract the navgroup section into a custom component to standardize the
styling accross the header
- [x] Add font sizing to the text
- [x] ~~Explore possiblity of having the border at the bottom move with
the opening. (for this I need to gitgud at animation and I am not right
now -- so will table it)~~
- [x] Update the links for each of the tabs to actually point to the
right location once @SeanKennyNF merges his code