All pathogens deselected now yields an empty map. (#204)
The problem: when a filter is empty, it defaults to yielding true. This
makes sense for most of the other filters, but for pathogen specifically
this is illogical. So now, if pathogen is empty, we return false for
that particular filter (since all the pathogens are preselected as
filters).

To check: go in and deselect all pathogens with any filter combo.