ESM maps (#248)
I just realized I did this whole thing in the wrong branch LMFAO but all
is well I think so

Anyway, in this PR I added in a new source + layer component that is
specific to the ESM's -- I am aware this takes away reusability of the
Pathogen Map component somewhat, but for the sake of speed that is ok.

I also added in a filter that lets you select the esm. This is a single
select filter and so I added in support for that as well.

Now when u select this it filters the data accordingly and then overlays
the esm on to the map.