106 add a filter for serotype on the dashboard (#205)
Adds serotype (which was on backend) to front end.

I also changed the way the pediatric age group filter works, resolving
#211 as well.

Also just a heads up on something a little bit tough to show: selecting
a pediatric age group can give the ILLUSION of growing one of the
pubbles (one goes from 92 to 108) but the total number of bubbles in the
area shrinks, and what actually happens is that the second bubble steals
from the first one because the first one is now smaller. Something to
keep note of. I verified (with a good amount of certainty) that the
filters are indeed actually reducing the number of records when used
together.

---------

Co-authored-by: Himanshu Ranka <himanshuranka2612@gmail.com>