Made the dashboards responsive (#253)
There are still many issues that will just take a lot more planning and
involevment to solve. Especially with the visualizations and data table.
Idk if I am just being silly and there is an easy way to solve it but
either way have thrown a little disclaimer on there.

It still kind of usable but we will need to make design changes across
the website and it will take more time and desiging. Let me know if
there is something that looks too bad and I have to fix it (other than
the viz cause they are actually asses and do not comply)