Fixed the bugs with incorrect data appearing for the clusters. (#331)
Resolves #269. The map flickers a little bit when the filters change
even if the cluster hasn't changed it's data. I think the intent of the
original caching code (which was causing this problem) was to remove the
flicker and to save generating elements we don't have to. This saves
generating elements we don't have to but still has problems with
flickering. Production also has problems with flickering, I think we
should do something more in-depth to remove that flickering but I'm just
fixing the bug for now.

![Peek 2024-05-27
18-24](https://github.com/serotracker/dashboard/assets/86806388/c332a563-b6ea-4928-901f-3132e52f7df6)