Allow the rows of the ArboTracker table to expand and reveal more information about the estimate. (#332)
Resolves #264. I'm not a CSS animation expert so I'm leaving the fancy
animation for another day. However, I added some very cool and fancy
stuff like the graph on the side. I really like the graph because that's
one graph that's a bit of a pain to generate and the data point
highlighted in green is the data point in the table in that graph which
is a touch I really love.

I'd like to make this change in SC2Tracker at some point but one thing
at a time.

![Peek 2024-05-29
14-10](https://github.com/serotracker/dashboard/assets/86806388/f15f843b-ce4f-44ca-9878-2714ee3fa5da)